Move middleware chain to middleware module
